refactor: update writer
